requirements.txt
pandas==1.5.3
GitPython==3.1.34